feat(#190): 종료 시간 계산 함수 분리
