+++ src/app/class/create/_components/ClassSchedule/ClassDay/TimeList.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import { ClearSVG } from '@/icons/svg';
 import { useClassScheduleStore } from '@/store';
+import { calculateEndTime } from '@/utils/parseUtils';
 
 interface TimeListProps {
   startTime: string;
@@ -47,35 +48,3 @@ const TimeList = ({ startTime, onChange, onRemove }: TimeListProps) => {
 };
 
 export default TimeList;
-
-// 종료시간 계산
-const calculateEndTime = (start: string, duration: number) => {
-  if (start.trim() === '') {
-    return '--:-- --';
-  }
-
-  const timeParts = start.split(':');
-  const dateObj = new Date();
-  dateObj.setHours(parseInt(timeParts[0]));
-  dateObj.setMinutes(parseInt(timeParts[1]));
-
-  // runningTime 은 분 단위
-  dateObj.setMinutes(dateObj.getMinutes() + duration);
-
-  let hours = dateObj.getHours();
-
-  const period = hours >= 12 ? 'PM' : 'AM';
-
-  if (hours > 12) {
-    hours -= 12;
-  } else if (hours === 0) {
-    // midnight
-    hours = 12;
-  }
-
-  // 항상 두 자릿수로 유지
-  const strHours = ('0' + hours).slice(-2);
-  const minutes = ('0' + dateObj.getMinutes()).slice(-2);
-
-  return `${strHours}:${minutes} ${period}`;
-};
